+++ packages/core/src/v3/apiClientManager/types.ts
@@ -1,0 +1,4 @@
+export type ApiClientConfiguration = {
+  baseURL?: string;
+  secretKey?: string;
+};