+++ packages/core/src/v3/index.ts
@@ -6,6 +6,7 @@ export * from "./limits";
 export * from "./logger-api";
 export * from "./runtime-api";
 export * from "./task-context-api";
+export * from "./apiClientManager-api";
 export * from "./schemas";
 export { SemanticInternalAttributes } from "./semanticInternalAttributes";
 export * from "./task-catalog-api";
@@ -28,6 +29,7 @@ export {
   flattenAttributes,
   primitiveValueOrflattenedAttributes,
   unflattenAttributes,
+  NULL_SENTINEL,
 } from "./utils/flattenAttributes";
 export { omit } from "./utils/omit";
 export {