+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @trigger.dev/nestjs
 
+## 3.0.0-beta.28
+
+### Patch Changes
+
+- Updated dependencies [6d9dfbc75]
+  - @trigger.dev/sdk@3.0.0-beta.28
+
+## 3.0.0-beta.27
+
+### Patch Changes
+
+- Updated dependencies [203e00208]
+  - @trigger.dev/sdk@3.0.0-beta.27
+
+## 3.0.0-beta.26
+
+### Patch Changes
+
+- @trigger.dev/sdk@3.0.0-beta.26
+
+## 3.0.0-beta.25
+
+### Patch Changes
+
+- @trigger.dev/sdk@3.0.0-beta.25
+
+## 3.0.0-beta.24
+
+### Patch Changes
+
+- @trigger.dev/sdk@3.0.0-beta.24
+
 ## 3.0.0-beta.23
 
 ### Patch Changes