+++ packages/otlp-importer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @trigger.dev/otlp-importer
 
+## 3.0.0-beta.28
+
+## 3.0.0-beta.27
+
+## 3.0.0-beta.26
+
+## 3.0.0-beta.25
+
+## 3.0.0-beta.24
+
 ## 3.0.0-beta.23
 
 ## 3.0.0-beta.22